Refactor huey calculation for display method

+++ Obstacle.pde
@@ -22,7 +22,7 @@ class Obstacle {
   }
 
   void display() { 
-    int huey =  360 * sin(frameCount) + 180; // (sin(frameCount/(360 * location.x * location.y)) + 2)
+    int huey =  int( 360 * sin(frameCount) + 180) ; // (sin(frameCount/(360 * location.x * location.y)) + 2)
     stroke(huey, 100, 100,100);
     fill(location.y,100,100,100);
     strokeWeight(4);
